Moves AuthenticationMethod to `Features` namespace, makes base class internal
