+++ src/Altinn.App.Core/Internal/Auth/AuthenticationTokenResolver.cs
@@ -1,10 +1,10 @@
 using Altinn.App.Core.Configuration;
+using Altinn.App.Core.Features;
 using Altinn.App.Core.Features.Auth;
 using Altinn.App.Core.Features.Maskinporten;
 using Altinn.App.Core.Helpers;
 using Altinn.App.Core.Internal.App;
 using Altinn.App.Core.Models;
-using Altinn.App.Core.Models.AuthenticationMethod;
 using Microsoft.Extensions.Options;
 
 namespace Altinn.App.Core.Internal.Auth;
