+++ test/Altinn.App.Core.Tests/Internal/Data/DataServiceTests.cs
@@ -1,8 +1,8 @@
 using System.Text.Json;
+using Altinn.App.Core.Features;
 using Altinn.App.Core.Internal.App;
 using Altinn.App.Core.Internal.Data;
 using Altinn.App.Core.Models;
-using Altinn.App.Core.Models.AuthenticationMethod;
 using Altinn.Platform.Storage.Interface.Models;
 using Moq;
