+++ test/Altinn.App.Core.Tests/Internal/Process/ExpressionsExclusiveGatewayTests.cs
@@ -10,7 +10,6 @@ using Altinn.App.Core.Internal.Instances;
 using Altinn.App.Core.Internal.Process;
 using Altinn.App.Core.Internal.Process.Elements;
 using Altinn.App.Core.Models;
-using Altinn.App.Core.Models.AuthenticationMethod;
 using Altinn.App.Core.Models.Process;
 using Altinn.App.Core.Tests.Internal.Process.TestData;
 using Altinn.Platform.Storage.Interface.Models;
